Backend/requirements.txt
boto3==1.39.13
boto3==1.39.12
botocore==1.39.13
botocore==1.39.12
fastapi==0.116.1
langchain_chroma==0.2.5
langchain_community==0.3.27
langchain_core==0.3.72
langchain_google_genai==2.1.8
langchain_text_splitters==0.3.9
langgraph==0.5.4
pydantic==2.11.7
pydantic_settings==2.10.1
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
supabase==2.17.0
